Infrastructure to create column stores
This introduces enough infrastructure to actually create the catalog
entries when the grammar has received COLUMN STORE clauses.

This creates pg_class entries (and relfilenodes) for the colstores, with
the corresponding pg_attribute entries; pg_cstore entries are also
created.  Add pg_depend entries too, so that everything is dropped when
the table is.

We create one relfilenode for each column store, which is referenced in
pg_class.relfilenode.  The column store AM is free to use this or not,
as well as create other relfilenodes by itself.  If we get a
pg_relfilenode catalog as discussed, this may be improved.

Implementation notes:

- Column store creation has two steps: DetermineColumnStores (called
  before the heap itself is created) examines table-level and
  column-level clauses, and merges both with column stores coming from
  inherited relations.  These are transformed into a list of
  ColumnStoreElem struct.

  The second step occurs after the heap itself is created, and records
  everything into pg_class, pg_cstore and pg_attribute entries,
  one relfilenodes for each column store, assigns OIDs to everything,
  etc.

- This patch creates new ObjectClass and ObjectType enum members for column
  stores, and associated fallout.

- Need to get rid of OIDs in pg_cstore, which requires some changes in
  the pg_depend entries and the deletion code introduced here.

- FIXME CloneColumnStores is a stub.  This function is used to recreate
  column stores when a table's heap is rewritten (cluster, vacuum full, etc).

- FIXME The event trigger system hasn't been updated to cope with the
  new object types.