relcache: don't consider nonzero pg_class.relam as an index
Right now, only indexes have the "relam" field set to a value different
from zero.  Once we introduce other kinds of access methods, that's no
longer the case.  In the one test that currently only checks for relam,
add a test for relkind also.  (Without this, creating a different kind
of relation with nonzero relam causes weird failures that the pg_index
row cannot be found).