Install temp hack to joinremovals
This aims to get around the problem where addRangeTableEntryForJoin() adds
all Vars which belong to both relations to the join RTE, thus causing the
expand_column_store_relations() function to add all column stores to the
table.

It would be nice if we could somehow only "expand" the column stores which
are actually needed, but at the time we call
expand_column_store_relations() we don't yet know which Vars are required
below the join. To properly fix this we'd need to somehow perform the
column store expansion after build_base_rel_tlists() has been called. This
seems hard, and requires more thought.

For now this just adds join removal code to get rid of the column store
joins when they're not used. The actual logic to determine "not used"
here is probably broken, but it's enough to get TPC-H working.

David Rowley