Allow column stores to have statistics
This adds a new API routine to the column store API which allows the
column store to gather a random set of tuples from the column store which
are then used by ANALYZE to generate statistics for the column store.

Please note that the tuples produced by the column store have no ability
to return the same tuples as were generated from the heap analyze. This
may cause problems if we ever allow multivariate statistics, however since
the statistics are stored under under the OID for the column store, rather
than the OID for the heap, then this might not be possible anyway.

This commit makes a change to update_attstats() to remove the OID to use
for the stat tuple's starelid. It seems that this parameter was redundant
anyway as we an just get the OID from stats->attr->attrelid. The new code
which performs the ANALYZE on the column store just switches out the
VacAttrStats->attr with the Form_pg_attribute belonging to the column
store. This means that we can just add/update all the stats in one call
to update_attstats(), rather than messing around calling it once for the
heap, then again for each column store.

David Rowley