Column store changes for nodeModifyTable & friends
Defines execColumnStore, with several methods
  a) ExecOpenColumnStores (~ ExecOpenIndices)
  b) ExecCloseColumnStores (~ ExecCloseIndices)
  c) ExecInsertColStoreTuples (~ ExecInsertIndexTuples)

Notes:

1) This disables HOT when there are column stores on the relation.
We may eventually fix this and relax the check so that we only disable
HOT when we update columns referenced in column stores (so we can allow
HOT on the heap part).

2) The inserts seem to be working just fine, but for updates this
complaints about missing ctid junk attribute.

3) Some of the signatures were simplified to use HeapTuple instead of
TupleTableSlot, but this seems a bad idea because everything else works
with TupleTableSlot (which also combine TupleDesc + HeapTuple).