planner: modify plan tree to include RTEs for column stores
This is intended as a proof of concept to see if this is a practical
idea for implementing column stores in PostgreSQL.

This code analyzes the parse tree before the simple_rel_arrays are
built, and it performs get_relation_info() on each relation in order to
determine if that relation has any column stores. For each relation in
the parse tree, and for each column store that these relation has, we
traverse the parse. Any Var which is found to belong to a column store
is replaced with a Var which points to the varattno of the column store
Var, instead of the heap Var.  We then add a JoinExpr in order to join
the heap to the colstore.  This means that column stores which are not
referenced are not added to the query. This perhaps needs more thought
for when it comes to the LEFT JOIN removal code, as I think this renders
that code useless.

At the moment the newly added column store JOIN, joins to the heap on
ctid, and to the column store on varattno = 1 (which has been hardcoded
to be the heap TID).

David Rowley