Allow replication slots to follow failover
Originally replication slots were unique to a single node and weren't
recorded in WAL or replicated. A logical decoding client couldn't follow
a physical standby failover and promotion because the promoted replica
didn't have the original master's slots. The replica may not have
retained all required WAL and there was no way to create a new logical
slot and rewind it back to the point the logical client had replayed to.

Failover slots lift this limitation by replicating slots consistently to
physical standbys, keeping them up to date and using them in WAL
retention calculations. This allows a logical decoding client to follow
a physical failover and promotion without losing its place in the change
stream.

A failover slot may only be created on a master server, as it must be
able to write WAL. This limitation may be lifted later.

This patch adds a new backup label entry 'MIN FAILOVER SLOT LSN' that,
if present, indicates the minimum LSN needed by any failover slot that
is present in the base backup. Backup tools should check for this entry
and ensure they retain all xlogs including and after that point. It also
changes the return value of pg_start_backup(), the BASE_BACKUP walsender
command, etc to report the minimum WAL required by any failover slot
if this is a lower LSN than the redo position so that base backups
contain the WAL required for slots to work.

pg_basebackup is also modified to copy the contents of pg_replslot.
Non-failover slots will now be removed during backend startup instead
of being omitted from the copy.

This patch does not add any user interface for failover slots. There's
no way to create them from SQL or from the walsender. That and the
documentation for failover slots are in the next patch in the series
so that this patch is entirely focused on the implementation.

Craig Ringer, based on a prototype by Simon Riggs

Backported to 9.4

Conflicts:
	contrib/pg_xlogdump/rmgrdesc.c
	src/backend/access/rmgrdesc/Makefile
	src/backend/access/transam/rmgr.c
	src/backend/replication/logical/decode.c
	src/backend/replication/slot.c
	src/include/access/rmgrlist.h
	src/include/replication/slot.h