Allow logical slots to follow timeline switches
Make logical replication slots timeline-aware, so replay can
continue from a historical timeline onto the server's current
timeline.

This is required to make failover slots possible and may also
be used by extensions that CreateReplicationSlot on a standby
and replay from that slot once the replica is promoted.

This does NOT add support for replaying from a logical slot on
a standby or for syncing slots to replicas.

(Backport from the 9.6 patch, allowing for the code move from
 logicalfuncs.c to xlogutils.c)