Logical decoding timeline following v2
The v1 timeline following logic failed to consider that a record could
be split across pages that are on different segments, where the new
segment contains the start of a new timeline. In that case the old
segment might be missing or renamed with a .partial suffix.

Rework the logic to be page-based and in the process simplify how the
last timeline for a segment is looked up.

This still relies on some ugly poking of globals in the walsender and
could use some refactoring love.

TODO: fixup backmerge the tests